+++ pysat/instruments/pysat_testing2d_xarray.py
@@ -37,7 +37,7 @@ clean = mm_test.clean
 preprocess = mm_test.preprocess
 
 
-def load(fnames, tag=None, inst_id=None, malformed_index=False,
+def load(fnames, tag='', inst_id='', malformed_index=False,
          start_time=None, num_samples=864, test_load_kwarg=None,
          max_latitude=90.):
     """Load the test files.
@@ -46,10 +46,12 @@ def load(fnames, tag=None, inst_id=None, malformed_index=False,
     ----------
     fnames : list
         List of filenames
-    tag : str or NoneType
-        Instrument tag (accepts '')
-    inst_id : str or NoneType
-        Instrument satellite ID (accepts '')
+    tag : str
+        Tag name used to identify particular data set to be loaded.
+        This input is nominally provided by pysat itself. (default='')
+    inst_id : str
+        Instrument ID used to identify particular data set to be loaded.
+        This input is nominally provided by pysat itself. (default='')
     malformed_index : bool
         If True, the time index will be non-unique and non-monotonic.
         (default=False)